agent/requirements.txt
web3==6.15.0
eth-account==0.11.0
python-dotenv==1.0.0
numpy==1.26.3
requests==2.31.0
aiohttp==3.9.1
scikit-learn==1.4.0
pandas==2.2.0
